+++ functions/MatchEnvironment.ts
@@ -1,4 +1,4 @@
-import { MatchFunction } from "$live/std/types.ts";
+import { MatchFunction } from "$live/types.ts";
 import { LiveState } from "$live/types.ts";
 
 export interface Props {
