--- std/ui/types/HTML.ts
@@ -1,4 +1,0 @@
-/**
- * @format html
- */
-export type HTML = string;
