--- std/ui/types/Image.ts
@@ -1,4 +1,0 @@
-/**
- * @format image-uri
- */
-export type Image = string;
