--- std/ui/types/Video.ts
@@ -1,4 +1,0 @@
-/**
- * @format video-uri
- */
-export type Video = string;
