+++ utils/unique.ts
@@ -1,0 +1,1 @@
+export const unique = <T>(array: T[]) => Array.from(new Set(array));
